Implement resume_job + job class hierarchy clean-up
